+++ CHANGELOG.md
@@ -7,7 +7,8 @@
 - change `jsx-no-cross-context-classes` options: before you have to set package name with option `libName`, e.g. `classname`. 
 So, you also have to import function only through default export and only with name `cn`.
 Now you have to set option `fn` with name of your function, no matter where from you've got it.
-For example, now you can set option like `fn: 'clsx'` and use named import like `import { clsx } from 'clsx'`.  
+For example, now you can set option like `fn: 'clsx'` and use named import like `import { clsx } from 'clsx'`.
+- restore removed `jsx-a11y` aribnb config 
 
 ## [0.12.0] - 2026-04-22